[AND-3984] RewardedInterstitial support. (#125)
